# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: 괌세이브카드 예약 관리 API
  description: |
    완전한 예약 파싱 이후 운영 관리 백엔드 API
    - OpenAI 기반 지능형 텍스트 파싱
    - 동적 필드 관리 (extras JSONB)
    - 감사 로그 및 변경 이력 추적
    - 실시간 SSE 이벤트
    - 일괄 작업 및 내보내기
  version: 1.0.0
  contact:
    name: 괌세이브카드 개발팀
    url: https://www.guamsavecard.com
servers:
  - url: https://guam-save-card-admin-production.up.railway.app
    description: Production server
  - url: http://localhost:3001
    description: Development server
security:
  - ApiKeyAuth: []
  - SessionAuth: []
paths:
...
  /api/bookings:
    get:
      summary: 예약 목록 조회
      description: 필터링, 검색, 페이징을 지원하는 예약 목록
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/ReviewFilter'
        - $ref: '#/components/parameters/ChannelFilter'
        - $ref: '#/components/parameters/PlatformFilter'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: 예약 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'

    post:
      summary: 예약 생성
      description: 새로운 예약을 수동으로 생성
      tags: [Bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: 예약 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/bookings/{id}:
    get:
      summary: 예약 상세 조회
      description: 특정 예약의 상세 정보, 감사 로그, 필드 정의 포함
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: 예약 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'

    patch:
      summary: 예약 수정
      description: 예약 정보 부분 수정 (낙관적 잠금 지원)
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - name: If-Unmodified-Since
          in: header
          schema:
            type: string
            format: date-time
          description: 낙관적 잠금을 위한 마지막 수정 시간
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookingRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '409':
          description: 충돌 (다른 사용자가 수정함)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'

    delete:
      summary: 예약 취소/삭제
      description: 예약을 소프트 삭제 (취소) 또는 하드 삭제
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: 삭제/취소 사유
                hard_delete:
                  type: boolean
                  default: false
                  description: 하드 삭제 여부
      responses:
        '200':
          description: 삭제/취소 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionAuth:
      type: apiKey
      in: cookie
      name: sessionId
  parameters:
    BookingId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: 예약 ID
    SearchQuery:
      name: q
      in: query
      schema:
        type: string
      description: 검색 키워드 (예약번호, 이름, 이메일)
    StatusFilter:
      name: status
      in: query
      schema:
        type: string
        enum: [pending, confirmed, cancelled, refunded, failed]
      description: 결제 상태 필터
    ReviewFilter:
      name: review
      in: query
      schema:
        type: string
        enum: [pending, needs_review, reviewed, confirmed, cancelled]
      description: 검수 상태 필터
    ChannelFilter:
      name: channel
      in: query
      schema:
        type: string
      description: 채널 필터
    PlatformFilter:
      name: platform
      in: query
      schema:
        type: string
      description: 플랫폼 필터
    DateFrom:
      name: from
      in: query
      schema:
        type: string
        format: date
      description: 시작 날짜 (YYYY-MM-DD)
    DateTo:
      name: to
      in: query
      schema:
        type: string
        format: date
      description: 종료 날짜 (YYYY-MM-DD)
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 페이지 번호
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: 페이지 크기
    Sort:
      name: sort
      in: query
      schema:
        type: string
        default: created_at
        enum: [id, reservation_number, korean_name, usage_date, total_amount, payment_status, review_status, created_at, updated_at]
      description: 정렬 필드
    Order:
      name: order
      in: query
      schema:
        type: string
        default: desc
        enum: [asc, desc]
      description: 정렬 순서
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        reservation_number:
          type: string
        confirmation_number:
          type: string
        channel:
          type: string
        platform_name:
          type: string
        product_name:
          type: string
        package_type:
          type: string
        total_amount:
          type: number
          format: float
        quantity:
          type: integer
        guest_count:
          type: integer
        korean_name:
          type: string
        english_first_name:
          type: string
        english_last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        kakao_id:
          type: string
        people_adult:
          type: integer
        people_child:
          type: integer
        people_infant:
          type: integer
        adult_unit_price:
          type: number
          format: float
        child_unit_price:
          type: number
          format: float
        usage_date:
          type: string
          format: date
        usage_time:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
        reservation_datetime:
          type: string
          format: date-time
        payment_status:
          type: string
          enum: [pending, confirmed, cancelled, refunded, failed]
        review_status:
          type: string
          enum: [pending, needs_review, reviewed, confirmed, cancelled]
        code_issued:
          type: boolean
        code_issued_at:
          type: string
          format: date-time
        memo:
          type: string
        extras:
          type: object
          description: 동적 필드 데이터 (JSONB)
        flags:
          type: object
          description: 데이터 품질 플래그
        origin_hash:
          type: string
        is_deleted:
          type: boolean
        lock_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        pagination:
          $ref: '#/components/schemas/Pagination'
        filters:
          type: object
    BookingDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            reservation:
              $ref: '#/components/schemas/Booking'
            field_definitions:
              type: array
              items:
                $ref: '#/components/schemas/FieldDef'
            audit_history:
              type: array
              items:
                $ref: '#/components/schemas/AuditRecord'
            metadata:
              type: object
    BookingResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Booking'
        validation:
          type: object
    CreateBookingRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/Booking'
      properties:
        _raw_text:
          type: string
          description: 원본 텍스트 (선택사항)
    UpdateBookingRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/Booking'
      properties:
        _reason:
          type: string
          description: 수정 사유
        _lock_version:
          type: integer
          description: 낙관적 잠금 버전
...
    FieldDef:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [string, number, date, time, datetime, boolean, select, multiselect, textarea, email, phone]
        required:
          type: boolean
        pattern:
          type: string
        options:
          type: object
        default_value:
          type: string
        placeholder:
          type: string
        help_text:
          type: string
        category:
          type: string
        sort_order:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AuditRecord:
      type: object
      properties:
        audit_id:
          type: integer
        booking_id:
          type: integer
        actor:
          type: string
        action:
          type: string
        diff:
          type: object
        previous_values:
          type: object
        current_values:
          type: object
        reason:
          type: string
        ip_address:
          type: string
        user_agent:
          type: string
        request_id:
          type: string
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_count:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error_code:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    ConflictError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
      properties:
        current_version:
          oneOf:
            - type: string
              format: date-time
            - type: integer
...